backend/app/services/simulation_service.py
from dateutil.relativedelta import relativedelta
from ..utils import to_float


def realizar_calculo_simulacao(dados):
    """
    SIMULAÇÃO SIMPLIFICADA PARA O PORTFÓLIO.
    A lógica de negócio original e proprietária foi removida.
    Esta função demonstra a estrutura de entrada e saída do cálculo.
    """
    # Exemplo de como acessar os dados que chegam
    consumo_total = 0
    for registro in dados.get("historico", []):
        consumo_total += registro.get("kWhProjPonta", 0)

    # Lógica de cálculo fictícia e simplificada
    custo_simulado_cativo = consumo_total * 0.85  # Ex: R$ 0,85 por kWh
    economia_simulada = custo_simulado_cativo * 0.15 # Ex: Economia de 15%

    # Monta um dicionário de resultados com a mesma estrutura que sua interface espera
    resultado_final = {
        "totais": { "custo_total_periodo": custo_simulado_cativo },
        "detalhes_mensais": [] # Pode gerar dados mensais fictícios aqui se precisar
    }
    return resultado_final